Make chart with number of nodes more accurate
The mettric of number of nodes is pulled from every ES node but it
does not contain an ES node label. This means that short variability
in number of nodes can be masked by metric pulled from different node
and go unnotticed if max() is used.

Using min() gives higher probability that the variablitiy will be
reflected in the chart.